various changes

+++ project/js/status.js
@@ -9,8 +9,14 @@ $(document).ready(function(){
 
      if(test==null){
          j_obj={
-             training1:[{"day":0,"time":"16.00-18.00","left":5,id:0},{"day":2,"time":"16.00-18.00","left":20,id:1},{"day":4,"time":"10.00-12.00","left":1,id:2},{"day":0,"time":"13.00-15.00","left":0,id:3}],
-             training2:[{"day":1,"time":"13.00-14.00","left":4,id:0},{"day":1,"time":"14.00-16.00","left":1,id:1},{"day":1,"time":"16.00-18.00","left":0,id:2},{"day":5,"time":"11.00-13.00","left":0,id:3}]
+             training1:[{"day":0,"time":"16.00-18.00","left":5,id:0},
+             {"day":2,"time":"16.00-18.00","left":20,id:1},
+             {"day":4,"time":"10.00-12.00","left":1,id:2},
+             {"day":0,"time":"13.00-15.00","left":0,id:3}],
+             training2:[{"day":1,"time":"13.00-14.00","left":4,id:0},
+             {"day":1,"time":"14.00-16.00","left":1,id:1},
+             {"day":1,"time":"16.00-18.00","left":0,id:2},
+             {"day":5,"time":"11.00-13.00","left":0,id:3}]
          }
          localStorage.setItem("schedule_data",JSON.stringify(j_obj));
          test=localStorage.getItem("schedule_data");
@@ -36,6 +42,8 @@ $(document).ready(function(){
      var data_j=JSON.parse(test);
 
     if(workout!=null && id!=null  && data_j[workout]!=null){
+        $("a[href='./schedule.html']").attr('href', './schedule.html?id_sched='+workout);
+
 
         var found=false;
         data_j[workout].forEach(element => {
@@ -51,12 +59,15 @@ $(document).ready(function(){
                         exists=true;
                     }
                 })
-                if(exists==false){
+                if(exists==false && element.left!=0){
                     $("#status_reserve").text("Uspešno ste rezervisali termin");    
                     user_reservations["reservation"].push({
                         "id":id,
                         "training":workout
                     });
+                    element.left-=1;
+                    
+
                     localStorage.setItem("user_data",JSON.stringify(user_reservations));
                 }
             }
@@ -71,5 +82,6 @@ $(document).ready(function(){
         $("#status_reserve").text("Došlo je do greške");
         $("#status_reserve").css("color","red");
     }
+    localStorage.setItem("schedule_data",JSON.stringify(data_j));
 
 });
